+++ src/python/dxpy/nextflow/ImageRef.py
@@ -118,6 +118,7 @@ class DockerImageRef(ImageRef):
             except subprocess.CalledProcessError:
                 err_exit("Failed to run a subprocess command: {}".format(cmd))
         # may need wait_on_close = True??
+        extracted_digest = self._digest
         if not self._digest:
             digest_cmd = "docker images --no-trunc --quiet {}".format(full_image_ref)
             extracted_digest = subprocess.check_output(digest_cmd, shell=True).decode().strip()
@@ -127,7 +128,7 @@ class DockerImageRef(ImageRef):
             folder=self._caching_dir,
             name=file_name,
             parents=True,
-            properties={"image_digest": self._digest}
+            properties={"image_digest": self._digest or extracted_digest}
         )
         return uploaded_dx_file.get_id()