+++ src/python/dxpy/nextflow/collect_images.py
@@ -44,10 +44,12 @@ def bundle_docker_images(image_refs):
     return bundled_depends
 
 
-def run_nextaur_collect(resources_dir):
+def run_nextaur_collect(resources_dir, nextflow_pipeline_params):
     """
         :param resources_dir: URL to the local(ized) NF pipeline in the app(let) resources.
         :type resources_dir: String
+        :param nextflow_pipeline_params: Custom Nextflow pipeline parameters
+        :type nextflow_pipeline_params: string
         :returns: Dict. Image references in the form of
             "process": String. Name of the process/task
             "repository": String. Repository (host) prefix
@@ -58,7 +60,9 @@ def run_nextaur_collect(resources_dir):
             "engine": String. Container engine.
         Runs nextaur:collect
         """
-    collect_cmd = "nextflow plugin nextaur:collect docker {}".format(resources_dir)
+    collect_cmd = "nextflow plugin nextaur:collect docker {} pipelineParams={}".format(
+        resources_dir, nextflow_pipeline_params
+    )
     _ = subprocess.check_output(collect_cmd, shell=True)
     with open(CONTAINERS_JSON, "r") as json_file:
         image_refs = json.load(json_file).get("processes", None)